Move feature extractor wrapper from init to a separate python file
